refactor: centralize music video api helpers
